+++ news/service/src/main/java/com/epam/lab/service/InterfaceService.java
@@ -2,17 +2,14 @@ package com.epam.lab.service;
 
 
 import com.epam.lab.dto.AbstractDTO;
-import com.epam.lab.model.Bean;
-
-import java.util.List;
 
 public interface InterfaceService <T extends AbstractDTO> {
 
     T create(T bean);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     T update(T bean);
 
-    T findById(int id);
+    T findById(long id);
 }
